label: stop using deprecated stuff in tests
Move 4 Linux tests that only test stuff from the selinux pkg
(TestDuplicateLabel, TestSELinuxNoLevel, TestSocketLabel, TestKeyLabel)
from label pkg to selinux, following commit 15c906f ("label: deprecate
selinux wrappers").

Remove some stub tests in label pkg (TestSocketLabel, TestKeyLabel,
TestProcessLabel, and parts of TestCheckLabelCompile) that already
have equivalents in selinux (TestSELinuxStubs).

Fix a few remaining issues in label to not use deprecated stuff
(ROMountLabel -> selinux.ROFileLabel, GenLabels -> InitLabels).
